+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Changed `Jackson2JsonBuilder(JsonFactory)` constructor visibility to package protected.
 - Added `Jackson2JsonBuilder.custom(JsonFactory)` to build a custom `Jackson2JsonBuilder`.
 - Changed all integration tests modules to use the `org.apiphany.tests` base package so that package protected stuff doesn't leak in.
+- Added Jackson 3 support (`jackson-databind` version `3.1.0`).
 
 ---